Remove Rust+iced gui in favor of a Vala+GTK for GNU/Linux
Will make a C#+WPF one for windows later, the UI for this is simple
enough, a cross plataform solution was giving me many headaches
which I don't think were worth it.

The main issue was iced's wgpu backend not playing nice with the old
Intel integraded GPUs I have on my Linux machines. I would do the UI
for Windows in Vala+GTK as well if the cross compilation process was
as simple as Rust's.